refactor: move rusqlite related code to clarity_src/vm/database/sqlite.rs
